requirements.txt
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
